In-website editing
The website now can handle editing a file when viewing it after a
search. Basex enforces the XML format is adhered to. Currently it is not
working with files that contain &. To complete the editing feature,
files with & need to work, the TEI schema needs to be enforced and it
needs to be in the browser file viewer as well.